+++ Tasks/UnityActivateLicense/UnityActivateLicenseV1/unity-activate-license.ts
@@ -15,7 +15,7 @@ async function run() {
         const unityExecutablePath = UnityPathTools.getUnityExecutableFullPath(unityEditorsPath, unityVersion);
 
         const logFilesDirectory = path.join(tl.getVariable('Build.Repository.LocalPath')!, 'Logs');
-        const logFilePath = path.join(logFilesDirectory, 'UnityActivationLog.log');
+        const logFilePath = path.join(logFilesDirectory, `UnityActivationLog_${Date.now()}.log`);
         tl.setVariable('logsOutputPath', logFilesDirectory);
 
         const unityCmd = tl.tool(unityExecutablePath)